requirements.txt
openai==2.30.0
instructor==1.14.5
pydantic==2.12.2
pydantic-settings==2.12.0
semgrep==1.156.0
bandit==1.9.4
PyYAML==6.0.3
pytest==9.0.2
pytest-timeout==2.4.0
puremagic==2.1.1
requests==2.32.4
